projects/ad579x_iio/scripts/requirements.txt
numpy==1.22.4
pyadi-iio==0.0.16
pylibiio==0.25
pynput==1.7.6
pyserial==3.5